Create README.md
This application was the final project of my C++ course. It simulates a parking-lot with entry and exit functionalities for vehicles and motorbikes. It keeps a track of the number of parking spots available in tandem with the entry and exit of vehicles. It accepts a text file with a list of motorbikes and cars., then lets the user choose whether to enter or exit a car or motorbike, shows details about the vehicle along with the parking spot.

What did I learn?

- Use of dynamic allocation of memory and using Valgrind to find leaks
- Encapsulation to increase modularity and reusability
- Utilising inheritance and polymorphism to write clean/reusable code
   - Abstract Base Classes
   - Virtual Functions

- Using stream objects effectively